Move response manager to its own file
