+++ src/dyn_message.h
@@ -25,13 +25,13 @@
 
 #include "dyn_core.h"
 #include "dyn_dnode_msg.h"
+#include "dyn_response_mgr.h"
 
 #define ALLOC_MSGS					  200000
 #define MIN_ALLOC_MSGS		     	  100000
 #define MAX_ALLOC_MSGS			      1000000
 
 #define MAX_ALLOWABLE_PROCESSED_MSGS  500
-#define MAX_REPLICAS_PER_DC           3
 
 typedef void (*func_msg_parse_t)(struct msg *);
 typedef rstatus_t (*func_msg_post_splitcopy_t)(struct msg *);
@@ -231,29 +231,6 @@ get_consistency_string(consistency_t cons)
 #define DEFAULT_WRITE_CONSISTENCY DC_ONE
 extern consistency_t g_write_consistency;
 extern consistency_t g_read_consistency;
-
-struct response_mgr {
-    bool        is_read;
-    bool        done;
-    /* we could use the dynamic array
-       here. But we have only 3 ASGs */
-    struct msg  *responses[MAX_REPLICAS_PER_DC];
-    uint8_t     good_responses;     // non-error responses received
-    uint8_t     max_responses;      // max responses expected
-    uint8_t     quorum_responses;   // responses expected to form a quorum
-    uint8_t     error_responses;    // error responses received
-    struct msg  *err_rsp;           // first error response
-    struct conn *conn;
-    struct msg *msg;
-};
-
-void init_response_mgr(struct response_mgr *rspmgr, struct msg*, bool is_read,
-                       uint8_t max_responses, struct conn *conn);
-// DN_OK if response was accepted
-rstatus_t rspmgr_submit_response(struct response_mgr *rspmgr, struct msg *rsp);
-bool rspmgr_is_done(struct response_mgr *rspmgr);
-struct msg* rspmgr_get_response(struct response_mgr *rspmgr);
-void rspmgr_free_response(struct response_mgr *rspmgr, struct msg *dont_free);
 
 struct msg {
     TAILQ_ENTRY(msg)     c_tqe;           /* link in client q */
